pyproject.toml
[tool.poetry]
name = "mxcube_video_streamer"
version = "1.3.0"
packages = [{include = "video_streamer"}]
license = "MIT"
description = "FastAPI Based video streamer"
authors = ["Marcus Oskarsson <oscarsso@esrf.fr>"]
maintainers = [
    "Marcus Oskarsson <oscarsso@esrf.fr>",
]
readme = "README.md"
homepage = "http://github.com/mxcube/fast-api-streamer"
repository = "http://github.com/mxcube/fast-api-streamer"
documentation = "http://github.com/mxcube/fast-api-streamer"
keywords = ["mxcube", "fast-api-streamer", "video_streamer", "mxcubeweb"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
pydantic = "^1.10.5"
fastapi = "^0.92.0"
pillow = "^9.4.0"
uvicorn = "^0.20.0"
pytango = { version = "==9.4.1"}
jinja2 = "^3.1.2"
websockets = "^10.4"
requests = "^2.32.3"


[tool.poetry.dev-dependencies]
pylint = [ { version = "==2.13.9", python = "<=3.7.1" }, { version = "2.15.3", python = ">=3.8" } ]
pre-commit = "2.20.0"
pytest = "7.1.3"
pytest-cov = "4.0.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
MJPEGDuoStreamer = 'video_streamer.tango.MJPEGDuoStreamer:main'
